Remove duplicated entries data from joinThread response
Summary:
In the near future, we want to call `joinThread` from `createThread`. In order to do so, we need to simplify `joinThread` and remove fetching that is duplicated there - entries and messages.
This diff removes calendar-fetching logic from `joinThread` for new clients, keeping it for old ones

Test Plan: Join thread and check if there are no errors. Check if entries are added to entryStore, check if data is not duplicated for new clients. Check old clients and verify that entries are still fetched correctly and visible in calendar when joining thread

Reviewers: palys-swm

Subscribers: ashoat, palys-swm, Adrian, atul

Differential Revision: https://phabricator.ashoat.com/D1044